% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/licensemanager_operations.R
\name{licensemanager_update_service_settings}
\alias{licensemanager_update_service_settings}
\title{Updates License Manager service settings}
\usage{
licensemanager_update_service_settings(S3BucketArn, SnsTopicArn,
  OrganizationConfiguration, EnableCrossAccountsDiscovery)
}
\arguments{
\item{S3BucketArn}{ARN of the Amazon S3 bucket where License Manager information is stored.}

\item{SnsTopicArn}{ARN of the Amazon SNS topic used for License Manager alerts.}

\item{OrganizationConfiguration}{Integrates AWS Organizations with License Manager for cross-account
discovery.}

\item{EnableCrossAccountsDiscovery}{Activates cross-account discovery.}
}
\description{
Updates License Manager service settings.
}
\section{Request syntax}{
\preformatted{licensemanager$update_service_settings(
  S3BucketArn = "string",
  SnsTopicArn = "string",
  OrganizationConfiguration = list(
    EnableIntegration = TRUE|FALSE
  ),
  EnableCrossAccountsDiscovery = TRUE|FALSE
)
}
}

\keyword{internal}
